+++ src/Encamina.Enmarcha.SemanticKernel.Plugins.QuestionAnswering/IKernelExtensions.cs
@@ -35,17 +35,35 @@ public static class IKernelExtensions
         Guard.IsNotNull(serviceProvider);
         Guard.IsNotNull(tokensLengthFunction);
 
-        var semanticTextMemory = serviceProvider.GetRequiredService<ISemanticTextMemory>();
-
         var semanticKernelOptions = serviceProvider.GetRequiredService<IOptionsMonitor<SemanticKernelOptions>>().CurrentValue;
         var modelName = semanticKernelOptions.CompletionsModelName ?? semanticKernelOptions.ChatModelName;
 
         var questionAnsweringPlugin = new QuestionAnsweringPlugin(kernel, modelName, tokensLengthFunction);
 
-        var memoryFunc = kernel.ImportMemoryPlugin(semanticTextMemory, tokensLengthFunction);
         var questionNativeFunc = kernel.ImportFunctions(questionAnsweringPlugin, PluginsInfo.QuestionAnsweringPlugin.Name);
         var questionSemanticFunc = kernel.ImportSemanticPluginsFromAssembly(Assembly.GetExecutingAssembly());
 
-        return memoryFunc.Union(questionNativeFunc).Union(questionSemanticFunc).ToDictionary(x => x.Key, x => x.Value);
+        return questionNativeFunc.Union(questionSemanticFunc).ToDictionary(x => x.Key, x => x.Value);
+    }
+
+    /// <summary>
+    /// Imports the «Question Answering» plugin and its functions into the kernel also adding the «Memory» plugin.
+    /// </summary>
+    /// <param name="kernel">The <see cref="IKernel"/> instance to add this plugin.</param>
+    /// <param name="serviceProvider">A <see cref="IServiceProvider"/> to resolve the dependencies.</param>
+    /// <param name="tokensLengthFunction">
+    /// A function to calculate the length by tokens of the chat messages. These functions are usually available in the «mix-in» interface <see cref="ILengthFunctions"/>.
+    /// </param>
+    /// <returns>A list of all the functions found in this plugin, indexed by function name.</returns>
+    /// <seealso href="https://en.wikipedia.org/wiki/Mixin"/>
+    public static IDictionary<string, ISKFunction> ImportQuestionAnsweringPluginWithMemory(this IKernel kernel, IServiceProvider serviceProvider, Func<string, int> tokensLengthFunction)
+    {
+        var semanticTextMemory = serviceProvider.GetRequiredService<ISemanticTextMemory>();
+
+        var questionAnsweringFunctions = kernel.ImportQuestionAnsweringPlugin(serviceProvider, tokensLengthFunction);
+
+        var memoryFunc = kernel.ImportMemoryPlugin(semanticTextMemory, tokensLengthFunction);
+
+        return memoryFunc.Union(questionAnsweringFunctions).ToDictionary(x => x.Key, x => x.Value);
     }
 }